Add attendance requirement to count for quorum
This commit adds an attendance requirement to the quorum rule.  By requiring a
voting member to be present for a certain number of meetings, it helps ensure
that quorum is met and decisions can move forward.  However, should a voting
member be absent, it helps ensure that decisions are not blocked.

In addition, it also establishes how a voting member can restore their voting
status.

Signed-off-by: Brian Warner <brian@bdwarner.com>