Update current TSC members and election process
This is a housekeeping change to update the results of the most recent election, and add some
explanation about the voting process. It also establishes a permanent location for the
self-nomination template.

It also migrates some administrative information out of the TSC charter text and into the README.
Charter text changes require a separate 2/3 vote, so having this info outside of the charter will
make it easier to update administrative information in the future.

Signed-off-by: Brian Warner <brian@bdwarner.com>