Allow expired seats to vote
The current membership is highly sensitive to movement of people and creates
voting dynamics that can be influenced by private interests.  By moving to
allow expired seats to retain their vote until replacements are found, the goal
is to create a more stable TSC with more continuity, while preserving voting
as a means of encouraging continued participation.